Requested changes

+++ Content.Shared/_MACRO/Species/Kodepiia/Consume/Components/ConsumedComponent.cs
@@ -1,13 +1,15 @@
 using Robust.Shared.GameStates;
 
 namespace Content.Shared._MACRO.Species.Kodepiia.Consume.Components;
-
+/// <summary>
+/// Entities with this component are considered "consumed" and track how many times they've been bitten.
+/// </summary>
 [RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
 public sealed partial class ConsumedComponent : Component
 {
     /// <summary>
-    /// Consumed value, added to whenever a consumer consumes the consumed.
+    /// How consumed this entity is, incremented by one every time they're consumed.
     /// </summary>
     [DataField, AutoNetworkedField]
-    public float ConsumedValue;
+    public int ConsumedValue;
 }
